add README and scripts for build and test
README.rst is work in progress and probably not proper rst format
yet but should still contain valuable information.

run_tests.sh is a script that runs unit and style tests on the
code. It can be used as a gate check in gerrit. Similar scripts
are standard practice in other open source projects.

JIRA: -
Change-Id: I5e586b346ff45f1151960a0e7fda2fe6072422c0
Signed-off-by: Hans Feldt <hans.feldt@ericsson.com>