Add support for action hooks in runner config
pre-start and post-stop intention is to be used to gather
information about the target system.

single-shot and periodic-action intention is to perform
actions on the infrastructure or cloud resources. For example
server live migration or network interface down.

Example of what can be added in the runner section:

    pre-start-action:
        command: "heat stack-show demo"
    periodic-action:
        interval: 10
        command: "ifconfig vboxnet1"
    single-shot-action:
        after: 30
        command: "nova show goofy.demo"
    post-stop-action:
        command: "nova list"

pre-start and post-stop data are added into the output file.
periodic and single-shot are not because that would interfere with
the actual sampled data. Besides the intention is not to log statistics
but do things with the infrastructure such as server live migration.

TODO: add sections to the output file, something like pre, data & post

JIRA: YARDSTICK-46
Change-Id: Ia059813fb74733f86368aea9c7a20e5afb71d228
Signed-off-by: Hans Feldt <hans.feldt@ericsson.com>