Add central output manager
Adds a static class, replicating current functionality
in a highly extensible module with a simple internal interface.

Having a centralised place to route all the output through makes
changes to the output format or destination, both dynamically as a
user option and statically as a design decision, much simpler and
easier.

Interface:

NOTE that all methods are called statically on the class,
not on initialised objects.

OutputMgr.init(cliargs):
Initializes the output manager and starts a (multiprocessing-safe)
queue for managing output. Calling write() before this step will
produce an error, since it (among other things) decides where output
should be directed during this stage.

OutputMgr.write(dict):
Writes the given object to the correct output. Accepts any python
object (dictionary) given to it, and currently dumps this as a JSON
object to the predefined output.

OutputMgr.all_done():
Signals the manager that all iteration loops have been completed,
allowing it to take appropriate actions and then shut down.

OutputMgr.shut_down():
Shuts down all sub-processes. This should be called any time
the output manager has been init()ialized, but may not have received
the all_done() call, such as in an "onexit" manager. It is safe from
overcalling, i.e. calling shut_down() on an already shut down, or
not yet initialised, output manager, will not result in an error.

Unimplemented or suggested extensions:

- Mode selection by command line
The current changes do not include changing or selecting modes. This
is probably best done by command line options (such as
--output-mode [ ONE_FILE | MULTI_FILES | STREAM ] ). The class is built
with feature in mind, including passing the command line arguments
to the initialiser, so the only things missing are the design decisions
on and implementation of how to act under a given mode, and the
CLI support for adding that flag, as well as a simple switch-like
code segment in the OutputMgr initializer.

- Network streaming
Actually a sub-issue of the above, streaming data to a data
analysis engine as it becomes available (immediately when it outputs)
instead of after the entire test is finished. This is a good option
to have, especially for very large test cases, where writing the data
to a file and then sending this file could cause a lot of overhead.

- Register a runner or scenario
While this system is decently flexible, it does not allow
for a single iteration of a single test file to print to multiple
targets. In order to add support for one-file-per-scenario or
one-file-per-runner modes, for example, there would be need for a
way to register the same with the handler, or rewriting to take a more
object-oriented approach, where the output from each scenario or
each runner is handled by a new instance of the manager.

JIRA: YARDSTICK-50
Change-Id: I0f2653876845a03f9fa99e5eb5fd9e67a3507801
Signed-off-by: raindirve <seanw@kth.se>