Adds extensible output manager - first stage
Feedback requested.

Adds a static class, fixing and replicating current functionality
in a simple internal interface.

Note that any new runners currently put()ting to Runner.queue would
need to have that line rewritten, as that functionality has been
migrated to this class.

Interface:

NOTE that all methods are called statically,
i.e. not on initialised objects.

OutputMgr.init(cliargs):
Initializes the output manager and starts a (multiprocessing-safe)
queue for managing output. Calling write() before this step will
produce an error, since it (among other things) decides where output
should be directed during this stage.

OutputMgr.write(dict):
Writes the given object to the correct output. Accepts any
dictionary given to it, and currently dumps this as a JSON object
to the predefined output.

OutputMgr.loop_done():
Signals the manager that a single iteration loop has been completed,
allowing it to take appropriate actions, such as printing a separator,
creating a new file to direct input to, or incrementing a counter.
Currently only increments a counter, but this behaviour is designed
to be easy to change.

OutputMgr.all_done():
Signals the manager that all iteration loops have been completed,
allowing it to take appropriate actions and then shut down.

OutputMgr.shut_down():
Shuts down all sub-processes. This should be called any time
the output manager has been init()ialized, but may not have received
the all_done() call, such as in an "onexit" manager. It is safe from
overcalling, i.e. calling shut_down() on an already shut down, or
not yet initialised, output manager, will not result in an error.

The only mode which is currently implemented is the single-file mode,
where all results are writted to a single output file
(defaulting to /tmp/yardstick.out and selectable by --output-file,
as per earlier functionality). Results from different iteration loops
are distinguished by a 1-indexed counter "loopnum" appended by the
write method.

The changes include adding this functionality to the task loop
in yardstick/cmd/commands/task.py

Integrated is the support for running multiple iterations of a given
test case, controlled by the --runs flag.

Example usage: yardstick task start --runs 8 samples/ping.yaml
would cause it to deploy the target stack, run the ping sample test
8 times, outputting the data to the default destination of
/tmp/yardstick.out, with the results from different
iterations (runs) being distinguished by the attribute
"loopnum": x (where x is a number between 1-8)

This also fixes the overwriting of serial scenarios bug (described
in JIRA: YARDSTICK-45).

Unimplemented or suggested extensions:

- Mode selection by command line
The current changes do not include changing or selecting modes. This
is probably best done by command line options (such as
--output-mode [ ONE_FILE | MULT_FILES | STREAM ] ). The
options in their entirety are already passed
to the initialiser, so the only things missing are the design decisions
on and implementation of how to act under a given mode, and the
CLI support for adding that flag, as well as a simple switch-like
code segment in the initializer.

- Network streaming
Actually a sub-issue of the above, streaming data to a data
analysis engine as it becomes available is an alluring alternative
to making one or several files to send after the test is
finished. Once a communication protocol has been established, this
could be easily added as one mode for output - only change necessary
beyond adding the option would be the implementation of the current
code stub _get_target() to decide according to the current mode
and/or the arguments given to it whether to return a writeable file
or a writeable socket.

- Register a runner or scenario
While this system is decently flexible, it does not allow
for a single iteration of a single test file to print to multiple
targets. In order to add support for one-file-per-scenario or
one-file-per-runner modes, for example, there would be need for a
way to register the same with the handler, or rewriting to take a more
object-oriented approach, where the output from each scenario or
each runner is handled by a new instance of the manager.

JIRA: YARDSTICK-50

Change-Id: I0f2653876845a03f9fa99e5eb5fd9e67a3507801
Signed-off-by: raindirve <seanw@kth.se>

AMENDED: Added copyright and license notice
AMEND2 2015-07-03 11:10 CEST: Fixed last-minute typo in runner/base
Signed-off-by: raindirve <seanw@kth.se>