Adds extensible output manager - first stage
Adds a static class, fixing and replicating current functionality
in a highly extensible module with a simple internal interface.

Having a centralised place to route all the output through makes
changes to the output format or destination, both dynamically as a
user option and as a design decision, much simpler and easier.

Also fixes the bug where serial scenarios would have their output
overwritten by subsequent scenarios in the same task file.

Interface:

NOTE that all methods are called statically,
i.e. not on initialised objects.

OutputMgr.init(cliargs):
Initializes the output manager and starts a (multiprocessing-safe)
queue for managing output. Calling write() before this step will
produce an error, since it (among other things) decides where output
should be directed during this stage.

OutputMgr.write(dict):
Writes the given object to the correct output. Accepts any
dictionary given to it, and currently dumps this as a JSON object
to the predefined output.

OutputMgr.all_done():
Signals the manager that all iteration loops have been completed,
allowing it to take appropriate actions and then shut down.

OutputMgr.shut_down():
Shuts down all sub-processes. This should be called any time
the output manager has been init()ialized, but may not have received
the all_done() call, such as in an "onexit" manager. It is safe from
overcalling, i.e. calling shut_down() on an already shut down, or
not yet initialised, output manager, will not result in an error.
It will simply do nothing at all.

Mode selection:

The only mode which is currently implemented is the single-file mode,
where all results are writted to a single output file
(defaulting to /tmp/yardstick.out and selectable by --output-file,
as per earlier functionality).

I recommend, but cannot make the one-sided design decision on, making
this a command-line option similar to --output-file, something like

yardstick task start --output-mode=MULTIPLE_FILES --runs 2 ./task.yaml

Unimplemented or suggested extensions:

- Mode selection by command line
The current changes do not include changing or selecting modes. This
is probably best done by command line options (such as
--output-mode [ ONE_FILE | MULT_FILES | STREAM ] ). The
options in their entirety are already passed
to the initialiser, so the only things missing are the design decisions
on and implementation of how to act under a given mode, and the
CLI support for adding that flag, as well as a simple switch-like
code segment in the initializer.

- Network streaming
Actually a sub-issue of the above, streaming data to a data
analysis engine as it becomes available is an alluring alternative
to making one or several files to send after the test is
finished. Once a communication protocol has been established, this
could be easily added as one mode for output - only change necessary
beyond adding the option would be implementing a function or other
switch, which picks an appropriate output destination depending on
what mode is active.

- Register a runner or scenario
While this system is decently flexible, it does not allow
for a single iteration of a single test file to print to multiple
targets. In order to add support for one-file-per-scenario or
one-file-per-runner modes, for example, there would be need for a
way to register the same with the handler, or rewriting to take a more
object-oriented approach, where the output from each scenario or
each runner is handled by a new instance of the manager.

(JIRA: YARDSTICK-45)
JIRA: YARDSTICK-50
Change-Id: I0f2653876845a03f9fa99e5eb5fd9e67a3507801
Signed-off-by: raindirve <seanw@kth.se>