Add support for multiple packet sizes in Pktgen
Pktgen Scenario support for multiple packet sizes defined in a list.
Supports both Arithmetic and Duration runner and is still compatible
with a test case with a single packet size which is not in a list.
Examples of valid packet sizes in the options section:

	packetsize:	packetsize: 60		packetsize:
	- 60					- 60
	- 124
	- 252

The data section in the output is structured as a JSON list object
if more than one packet size is provided.

JIRA: YARDSTICK-47

Change-Id: I819fe82175141cd069d5f2f431feb6fa010cc8c9
Signed-off-by: Kristian Hunt <kristian.hunt@gmail.com>