Add new stepper runner
The "stepper" runner will use an input value of the test for every
run from a pre-defined list from the task file.
Example runner section of the task file:
    type: Stepper
    scenario_option_name: number_of_ports
    steps:
    - 10
    - 15
    - 20

If the value of the scenario option defined in the options section
is different from the first value of the steps list, then scenario
option value is added to the front of the list.

JIRA: YARDSTICK-47

Change-Id: I5bde9b78cb356499c338ef3da26ac1783670887d
Signed-off-by: Kristian Hunt <kristian.hunt@gmail.com>