Update patch set 4
Patch Set 4:

(1 comment)

> Can you please describe what value this changeset adds?
 > What is the difference in functionality compared to what we have
 > right now?

Being able to run several runs is worthless without being able to see the output. This is why both are bundled into the same change. Several runs would just overwrite each other.

Also fixes the YARDSTICK-45 bug since it reworks the system that bugs up. 

Beyond that, it's designed to be easily extendable - this is why there's several multiple-choice entries that are currently single-choice. Intended that all that's needed is to add the mode and the rest will be easy, instead of needing to implement the rest of an already planned out feature then and there.

Patch-set: 4
Label: Verified=0