Add unit test for ssh
Unit test file for ssh.py originates from rally. It is modified to
be consistent with the ssh module used in yardstick - some tests for
missing functionality are removed and naming convention is changed
(ssh in yardstick vs sshutils in rally).

Running of unittest from run_test.sh is NOT enabled.

JIRA:-

Change-Id: I3ea554e3fbe6f1f906778f88ddd859424331c8df
Signed-off-by: Kristian Hunt <kristian.hunt@gmail.com>