Add unit test for logical model
Add unit tests for logical model (model.py). This test suite does not
include tests for unfinished functionality - some parts of methods or
in some cases whole methods with unresolved TODOs are not tested.

This test suite is not complete, for most methods only the default
case is tested.

Running of unittest from run_test.sh is NOT enabled.

JIRA:-

Change-Id: I0984fb3655513f5ad67dcc7436461990da9296e7
Signed-off-by: Kristian Hunt <kristian.hunt@gmail.com>