Add script for running yardstick tasks back-to-back
This script enables to run multiple yardstick tasks in the yardstick
daily build without having to change the Jenkins job in the releng
repository.

This script should be executed from the main yardstick directory just
as run_tests.sh.

If at least one of the tasks listed returns a non-zero value then the
script continues executing, but in the end will exit with value 1.

JIRA: YARDSTICK-106

Change-Id: I8a6bb7e6c03ec551709ff66a45aad7257fb36e92
Signed-off-by: Kristian Hunt <kristian.hunt@gmail.com>