Update patch set 1
Patch Set 1:

Just a question; I suppose the exit code of yardstick if one of the tasks fails is non-zero, causing the script to exit at that point, skipping running the remaining tasks. If so can we make sure this doesn't happen so we can run all the tasks.

Patch-set: 1