Update patch set 2
Patch Set 2:

I planed Dispatcher to be the backend of OutputMgr and currently focus on result storage part. There several ways to store result local file, remote server or local db, ways to handle dispatching this stuffs is needed, so dispatcher is introduced.
We need find a way to eliminate Dispatcher and OutputMgr functionality confusion.

Patch-set: 2