Add test suite functionality to task commands
This patch proposes a more permanent solution to replace the already
existing one (YARDSTICK-106).

A test suite is a .yaml file which consists of multiple task files which
are meant to run as a one set. Each test case in the test suite has to
specify the location of the test .yaml task file and can specify additional
task arguments or a task arguments file.

Two command line arguments are added to task command.
First command is --suite which indicates that the intention is to run
a test suite and the second is --test-cases-dir which points to the directory
where the test .yaml files are located.
If --test-cases-dir is unspecified then it defaults to tests/OPNFV/test_cases/
according to https://etherpad.opnfv.org/p/yardstick_folders

JIRA: YARDSTICK-107

Change-Id: I69508b3b536dd592b6240a2a32dd5d665e0ac4b4
Signed-off-by: Kristian Hunt <kristian.hunt@gmail.com>