Add test suite functionality to task commands
This patch proposes a more permanent solution to replace the already
existing one (YARDSTICK-106).

A test suite is a .yaml file which consists of multiple task files which
are meant to run as a one set. Each test case in the test suite has to
specify the location of the test .yaml task file and can specify additional
task arguments or a task arguments file.

It is possible to specify the directory where the task files are located
in the suite file. If it is left unspecified then it defaults to
tests/opnfv/test_cases according to https://etherpad.opnfv.org/p/yardstick_folders

A Command line argument --suite is added to the task command. This
argument indicates that the intention is to run a test suite.

JIRA: YARDSTICK-107

Change-Id: I69508b3b536dd592b6240a2a32dd5d665e0ac4b4
Signed-off-by: Kristian Hunt <kristian.hunt@gmail.com>