Update patch set 8
Patch Set 8:

(1 comment)

Yes a jenkins/gerrit job will eventually rebuild the docker image and push it to the opnfv docker hub account if there are changes. It is not clear at the moment exactly how this will be done so to start with the docker image will be pushed manually.

The jenkins ci job will always pull the latest docker image from the hub and run the test,

Patch-set: 8