Heat context code refactor
Heat context code refactor to cater for the evolution of the
Yardstick framework.

This is the initial version, very rough with many TODOs in the code,
but show the idea of how the new yardstick context will look like.

In test_case.yaml context segment will add "context_type" to indicate
the context type, default is Heat.

Welcome your review and comments.

JIRA: YARDSTICK-168

Change-Id: Ida0ce12c17cd9b88d7acfb4c9eb1ac6986394b38
Signed-off-by: QiLiang <liangqi1@huawei.com>