Update patch set 2
Patch Set 2:

I would say you need to update the image everytime when:
1) The Dockerfile has changed
2) The first script that you call in the docker has changes

If yardstick-verify is that script, then I would say so.

Anyway, a good idea also is to create a jenkins daily job to build an image everyday. The good think of docker is that if there are not changes in the image, you dont need to spend more storage in dockerhub. It works with layers. So, there is no much overhead by doing it everyday.

Patch-set: 2