Update patch set 2
Patch Set 2:

The image only have to be rebuilt and uploaded to docker hub if the Dockerfile or run_benchmarks have changed. yardstick-verify does not impact the image as it is the second script called.

Yes we should look into how we can automatically rebuild the Docker image in Jenkins. Is functest already doing that ?

Patch-set: 2