Remove etc and source folder from yardstick documentation
conf.py and logo will come from releng which is common for all projects.
The directory structure in docs folder has not been decided yet so this
change does not require you to change your current way of working.

Once this directory structure is decided by opnfvdocs and releng, the changes
need to be reflected back to all projects, including yardstick.

Change-Id: If0558474a819f4a44e34f790d26c5486751b8266
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>