Adapt documentation to latest guidelines
Adapt Yardstick documentation according to the latest guidelines from the
Releng project.

The only change in test case description TC001 and TC002 is to remove logo.
The same for the test case description v2 template.

The previous version of test case description is deleted from repo.

This commit also adds index for docs/yardstick and for
docs/user_guides/framework.

Change-Id: Ib6480c3c29ce0f0ee22c9ad717439f6a293ab842
Signed-off-by: Ana C <ana.cunha@ericsson.com>