Changes Arithmetic runner to accept several step parameters
The Arithmetic runner is modified to accept several named
parameters to iterate over in each scenario. Possible to chose
either to iterate through them in a nested-for-loop fashion
(default) or in a tupled loop fashion (combine i:th element
of each input). Still possible to only provide one parameter
to step/iterate over, as before.

All existing yaml files with Arithmetic types have been modified
to run the tests as before, and to provide same results as before
this change. The only change is in sample file pktgen.yaml to
provide a valid example of how to use this modification.

It is also possible to step backwards, i.e. start > stop, with
a negative step parameter value.

(Noticed that tests are terminated when there is SLA error
problems in "monitor" mode. Should be fixed in another commit)

Change-Id: Ib2ebd24b71dd55d6817cee8e67026a0dd13a9e17
JIRA: -