Update patch set 3
Patch Set 3:

Hi Qi Liang,
I had a look and have a question for you.

I think this is fine with me and if I correctly understood, this means I will need to add to the list my test cases and implement a function that converts the result into a valid influxdb string for each TC.

I could not have 100% understanding of the "behind the scene" of Yardstick architecture, so my following observation could be a stupid one. 
An alternative to this mechanism could be that, regardless the specific test case sending results, the dispatcher could just convert the dictionary into a string using all its keys().

(see key1=val1,key2=val2)

This would make life easier to a yardstick developer that wants to develop a new test case in the future. 
Does that makes sense to you? Do you have any thoughts on this?

Patch-set: 3