Update patch set 3
Patch Set 3:

Hi,

I think your suggestion having generic format func is a great idea.

For simple outputs (e.g. Ping) 
   
      {"key1": val1, "key2": val2}

we could simply generate:

      "key1=val1,key2=val2"


For nested dictionaries (e.g. cpuload)

      { 
            "mpstat": {
                "cpu0": {
                    "%sys": "0.00",
                    "%idle": "99.00"
                } 
             }
      }

The results can be flattened:

      "mpstat.cpu0.sys=0.00,mpstat.cpu0.idle=99.00"


There is also a case in cpuload with the dictionary values
being lists:

      { 
            "loadavg": [
                "1.09",
                "0.29",
                "0.10",
                "1/84",
                "1293"
            ]
      }

This could maybe be resolved to
  
      "loadavg0=1.09,loadavg1=0.29,..."   

or we can update cpuload scenario to not dump list structures in the result.)

Patch-set: 3