Update patch set 3
Patch Set 3:

Do we need more tags in the result output ?

Im a bit novice on InfluxDB but I assume that tags can be used for searching and selection of test results? 

- runner_id ?

It is already available in runner/scenario_cfg and may be needed as a tag if we want to search for results belonging to the same runner.

- test case name ?  

Allows for searching results using tc name as index.

Could be the same as the (base)name of the yaml file. 

   E.g. tc=opnfv_yardstick_tc001

- task_id ?   

   We maybe should generate a 'task_id' so runner results belonging to the same task file run can be selected?
      
   E.g add scenario_cfg["task_id"] = os.pid()  (in task.py) and use as tag.

( maybe instead of using os.pid() we should generate a uuids )

Patch-set: 3