Update patch set 2
Patch Set 2:

Looks Ok, but something to ponder.

With this patch it should be possible to run tests "locally" without having to boot a VM instance for test execution:

https://gerrit.opnfv.org/gerrit/#/c/5385/

This would of course require that the packages are available locally, i.e. either installed/cloned by the scenario itself,  or installed with yardstick setup.py.

Patch-set: 2
Label: Verified=0