support os+odl scenario for ipv6
As we known, the network of os+odl scenario is still not available.
so this patch still need to do more debug work.
however, Jan 19th is code freeze day, i think a good way is to merge it at first
when os+odl become available, i will do rest debug work and fix it

Change-Id: I518f759069e74d2e57e40b52dfecaef4b4d26a02
Signed-off-by: kubi <jean.gaoliang@huawei.com>