add precondition check
as we discussed yersterday, for daily jenkins task,
i have a new idea, i add a precondition check and a key parameter in test case,
if environment info(eg. "DEPLOY_SCENARIO") meet the preconditon which was defined in test case ,
this test case will run, if not meet, this test case will skip.

and default is allow all test case to run, so this patch will not influence existing test case.
any comments are welcomed

Change-Id: I4300ac58994d51c0ddb4dd6d58b7191f796ddcee
Signed-off-by: kubi <jean.gaoliang@huawei.com>