Yardstick input for Release central doc
This change adds a file, userguide.rst, that will be used by the release
central document as a summary of the Yardstick Project.

This file shall not be included in the index of the userguide.

The file has been renamed to testusage.rst, as per request.

Change-Id: Id8099349a7e626be3d800499bf98171d23239ded
Signed-off-by: Ana C <ana.cunha@ericsson.com>
(cherry picked from commit 3bcd8816da2d38f5c0d3c4358a47aac2ac19f873)