iperf3: fix for influxDb schema errors
Convert integers to float before dumping to influxDb.

Iperf3 is not always consistent about data types in its
JSON output. This may cause schema errors when values with
different types (integer and float) are written to the same
field in InfluxDb.

This is a quick fix. A Future improvement would be to write a
format func for the data.

Change-Id: I7011adc19693f91db5889554fd6bec0d46dc6679
Signed-off-by: Jo¶rgen Karlsson <jorgen.w.karlsson@ericsson.com>