Introduction of a new environment variable indicating a virtual environment
This patch introduces a new environment variable: DEPLOY_TYPE [bm | virt],
indicating if the opnfv stack is deployed bare-metal or virtualized.
This can be used for several purposes, but in this patch it is used
to tell "fetch_os_creds.sh" wheather it is a BM or virt deployment.
Also see: https://gerrit.opnfv.org/gerrit/#/c/11541/
NOT VERIFIED
CAN MERGE AS SOON AS VERIFIED

Change-Id: I2d4385dc9a2786ce51e44081967a77736276696e
Signed-off-by: Jonas Bjurel <jonas.bjurel@ericsson.com>