Introduction of a new environment variable indicating a virtual environment
This patch introduces a new environment variable: DEPLOY_TYPE [bm | virt],
indicating if the opnfv stack is deployed bare-metal or virtualized.
This can be used for several purposes, but in this patch it is used
to generalize the invocation of exceptions needed for virtualized deployments
without having to hardcode POD names being virtualized.

Current exemption: if [ "$INSTALLER_TYPE" == "fuel" ] \
                      && [ "$DEPLOYMENT_TYPE" == "virt" ]; then
                     <modify openrc>

NOTE: This change is intended to be fully backwards compatible.

NOT VERIFIED
CAN MERGE AS SOON AS VERIFIED

Change-Id: I2d4385dc9a2786ce51e44081967a77736276696e
Signed-off-by: Jonas Bjurel <jonas.bjurel@ericsson.com>