Update patch set 1
Patch Set 1:

I have 2 questions regarding this change.

First one is about having node name in this script.

Is the change internalURL -> publicURL only needed for ericsson-virtual1 or it is valid for all the fuel virtual deployments no matter where/which machine it is done?

If so, can node name be removed from here.

Apart from this, should this change be in here or in the script below that is used for extracting openrc because yardstick also uses openrc extracted by the below script. So fixing the source could perhaps be better rather than fixing it only for yardstick.

https://gerrit.opnfv.org/gerrit/gitweb?p=releng.git;a=blob;f=utils/fetch_os_creds.sh

Patch-set: 1