ubuntu_cloud_img: Allow pwd console access
Currently if there is a port bind failure it is impossible to access a
VM based on Ubuntu server cloud image to debug.

This change allows access on the console for user ubuntu by was of a
password that is printed to the console during start up; The same way as
VMs based on the cirros image do.

Change-Id: Iac242ba1f3f14ae77a7b00cf2befd6fa4b65af09
Signed-off-by: Billy O'Mahony <billy.o.mahony@intel.com>
Reviewed-by: Ana Cunha <ana.cunha@ericsson.com>
Reviewed-by: Jorgen Karlsson <jorgen.w.karlsson@ericsson.com>
Reviewed-by: Jonas Bjurel <jonas.bjurel@ericsson.com>
Reviewed-by: Daniel Smith <daniel.smith@ericsson.com>
Reviewed-by: Mark D. Gray <mark.d.gray@intel.com>