[WIP][Yardstick-233]latency measurment by using pktgen-dpdk
VM A runs DPDK pktgen and is connected back-to-back with
VM B running testpmd for forwarding packets.

Because the Lua side of using the Latency is not completed
only the command line. So I use linux expect to setup test env
and get latency temporarily.

When the Lua API of Latency is supported, I will turn to Lua API.

Besides, some code still need to test.

JIRA:Yardstick-233

Change-Id: I90ae64b3d198c79f159275ae35715eb284f05080
Signed-off-by: wu.zhihui <wu.zhihui1@zte.com.cn>