Update patch set 2
Patch Set 2:

(5 comments)

Thanks for the effort, we, kvm4nfv project is looking for such test case and glad to find this effort.

I leave some comments for the implementation, and have one basic question is, why do we do the test on the yardstick, instead of on the vsperf? I think vsperf has some infrastructure support already.

I'm not against to do this on yardstick(in fact, I'm really glad to see this effort), just want to know some background information.

Patch-set: 2
Label: Verified=0