ARM64: Add suppoort for arm64 architecture
Added Support for arm64 architecture so that yardstick framework
can be ran on arm-pod1 the armband CI pod. The modifications include
Testing for arm64 architecture an running arm64 specific code.
The Openstack doesn't boot 3.13 kernel so it was modified with 3.19
kernel available in vivid cloud image of ubuntu jira case ARMBAND-40.

Change-Id: I906d4adbcf2ec2c0f1d695d9916ede42f4151080
Signed-off-by: Paul Vaduva <Paul.Vaduva@enea.com>