ARM64: Add support for aarch64 architecture
Added support for arm64 architecture so that yardstick framework
can be ran on arm-pod1 the armband CI pod. The modifications include
testing for presence of arm64 architecture nodes in the pod and running
arm64 specific code (building yardstick server image on arm64 version of
ubuntu cloud.
Openstack doesn't boot 3.13 kernel(arm64) so it was replaced with 3.19
kernel(also arm64), available for vivid cloud image of ubuntu.
The Dockerfile for building the ci image was also mofied to add necessary
tools for cross-building an arm64 image on amd64 host (the jumpserver)
The jira case for this commit is ARMBAND-40.

Change-Id: I906d4adbcf2ec2c0f1d695d9916ede42f4151080
Signed-off-by: Paul Vaduva <Paul.Vaduva@enea.com>