Update patch set 1
Patch Set 1:

Ok. One question. When I define the pod.xml, I have to provide the user/password to access the nodes, but the problem is that the nodes are not accessible from the jumphost because of a public key. They are only accessible from Fuel VM... that could be a constraint..

Patch-set: 1
Reviewer: Gerrit User 459 <459@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0