Update patch set 2
Patch Set 2:

The ubuntu image (trusty-server-cloudimg-amd64-disk1.img) we previously added in dockerfile is only used for storperf. We only use amd version. 

The base ubuntu image used for building yardstick-trusty-server can be either arm64 version or amd64 version, this base image will be downloaded every CI job. There is a "download" function in yardstick-image-modify to do it.

So the ubuntu image here is

Patch-set: 2