Update patch set 3
Patch Set 3:

The key files generated is for per test case.
We plan to support concurrent in the feature, which means there will be more than one pair of keys. So I think the concurrent support work and key removal will be done in the future. So far I think it is good enough, as the old keys have been removed from the repository and the new keys are only in local.

Patch-set: 3
Label: Verified=0