Update patch set 4
Patch Set 4:

The trick it seems is to clear the handlers and create them only once, then we need to hold a reference to the handlers to we can adjust the loglevel on each handler.

We could create the handlers in __init__.py and only set the loglevels in cli.py or in the rest api.

Patch-set: 4
Label: Verified=0