Update patch set 4
Patch Set 4:

hi, Jing, thanks for update, But I'm not sure if it is great for Readme, It seems that most new info is the copy of UserGuide. I guess a brief and simple guide will be better. It could mainly include the Command line of every step.  and the userguide link could be added to let the user to check the detail infomation.

Patch-set: 4
Label: Verified=0