Update patch set 7
Patch Set 7: Code-Review-1

It seems me this README.rst is not necessary, since we already have the user-guide for people to know how to install yardstick. Also README.rst can not be output to the artifact, so it doesn't make sense to use this format. Shall we remove this README to avoid duplication?

Patch-set: 7