heat: fix key_uuid format
The .format() won't work, since it can't slice the uuid

We have to convert the uuid to string before we slice it.
I thought .format() would implicitly call __str__() before
applying width, but that is not the case.

'files/yardstick_key-{:.{width}}'.format(self.key_uuid, width=8))

We also need to define a constant short uuid lenght, we can't
hardcode the length to 8 everywhere.

Create a helper function to standardize the generation
of the short key uuid and use that helper function everywhere

Change-Id: I59e051bfe697587e967f93f5b8f209e0e7daa5c7
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>