uwsgi: move init_db to uwsgi entry point
uwsgi will call the 'callable' function
which should normally be Flask.__call__().

But we need to init the db first, so make a wrapper
function that inits the db and then calls app()

Also refactor add_resource into a function and use
for loop instead of reduce.  reduce is not really
approriate here since we aren't consuming the return
value

JIRA: YARDSTICK-543

Change-Id: I692d6d42de09f7d6ecf1a67a22e3019d97a4f3ca
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>